refactor: cleanup UI components (#21)
* refactor: cleanup UI components

Use a better foldering structure for reusable components.

* refactor: use named exports on icons

* fix: use proper element tags for layout components

* refactor: rename icons folder
