+++ components/helpers/NextChakraLink.tsx
@@ -5,10 +5,12 @@ import { Link as ChakraLink, LinkProps as ChakraLinkProps } from '@chakra-ui/cor
 
 export type NextChakraLinkProps = PropsWithChildren<NextLinkProps & Omit<ChakraLinkProps, 'as'>>
 
-export function NextChakraLink({ href, as, replace, scroll, shallow, prefetch, children, ...chakraProps }: NextChakraLinkProps) {
+function NextChakraLink({ href, as, replace, scroll, shallow, prefetch, children, ...chakraProps }: NextChakraLinkProps) {
   return (
     <NextLink passHref href={href} as={as} replace={replace} scroll={scroll} shallow={shallow} prefetch={prefetch}>
       <ChakraLink {...chakraProps}>{children}</ChakraLink>
     </NextLink>
   )
 }
+
+export default NextChakraLink
