+++ components/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as Logo } from './Logo'
+export * from './Logo'
